Terminando relacionamento de venda com pagamento, e criando nova entidade para estoque.

+++ src/main/java/br/com/supermercado/api/models/Pagamento.java
@@ -16,7 +16,7 @@ public class Pagamento {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_pagamento")
-    private int idPagamento;
+    private Long idPagamento;
 
     @Column(name = "data_pagamento")
     private String dataPagamento;
